Zombie killer
A bug in the KoduService error handling code allowed a thread to stay alive indefinitely.  This in turn kept the Mutex alive which we use to prevent multiple copies of Kodu from being launched.  So if you got a Zombie Kodu running, it would prevent any other Kodu startup until you rebooted.  This change detects this case, kills the zombie and then runs.